rebasing with main and changing project layout
